Refactor: 파일 확장명 tsx로 변경
