+++ src/features/main/views/Main.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import MainElem from "./MainElem";
 import Data from "../../../data.tsx";
 import { MainContainer } from "../styles/MainStyle.tsx";
-import type { DataType } from "../../../data.tsx"; // Import the DataType interface
+import { DataType } from "../../../data.tsx"; // Import the DataType interface
 
 function Main() {
   const sortedData = [...Data].sort((a: DataType, b: DataType) => new Date(b.uploadDate).getTime() - new Date(a.uploadDate).getTime()); // date 기준으로 오름차순 정렬
