+++ src/features/shorts_btns/styles/ShortsBtnsStyle.tsx
