+++ src/pages/shorts/styles/ShortsStyle.tsx
