+++ src/typings.d.ts
@@ -1,0 +1,12 @@
+// declare module "*.png" {
+//   const value: any;
+//   export = value;
+// }
+
+// declare module "*.jpeg" {
+//   const value: any;
+//   export = value;
+// }
+
+declare module "*.png";
+declare module "*.jpeg";
