+++ tsconfig.json
@@ -106,5 +106,5 @@
     // "skipDefaultLibCheck": true,                      /* Skip type checking .d.ts files that are included with TypeScript. */
     "skipLibCheck": true /* Skip type checking all .d.ts files. */
   },
-  "include": ["typings.d.ts"]
+  "include": ["src/typings.d.ts"]
 }
