--- typings.d.ts
@@ -1,4 +1,0 @@
-declare module "*.png" {
-  const value: any;
-  export = value;
-}
